+++ api/src/banner/getImage.js
